pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "llama-bench"
version = "0.1.0"
description = "Benchmarking and optimizer CLI for llama.cpp llama-server"
requires-python = ">=3.10"
dependencies = [
    "click>=8.0",
    "rich>=13.0",
    "requests>=2.28",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = ["pytest>=7.0", "pytest-mock>=3.10"]

[project.scripts]
llama-bench = "llama_bench.cli:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["llama_bench*"]
